requirements.txt
requests
beautifulsoup4
pyyaml
python-dateutil
pytz
filelock
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
